+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 # [Unreleased]
 ### Added
 - Analytics usage may now be switched on/off with an environment variable
+- Old-style links to image details are now supported.
 
 ### Fixed
 - Logos disappearing when user entered shopping cart and checkout views.